Enabled weakrefs in game contexts again and provided independent randomizer for `core.context` module.
The former was made to ensure that the game context is appropriately unregistered
when we lose any other references to the game controller, and to avoid cyclic
references.
The latter ensures that generating unique context identificators is not affected
by global random seeding (that caused problems with registering/unregistering
contexts in test cases with fixed random states).